NHA-168
Lots of files changed due to readability changes throughout the project. Some files changed due to line ending changes.
Build pipeline created, publishes artifact on master and dev branches.

+++ tjeneste/src/main/java/no/arkivverket/helsearkiv/nhareg/util/FanearkidValiderer.java
@@ -10,18 +10,18 @@ import no.arkivverket.helsearkiv.nhareg.tjeneste.KonfigparamTjeneste;
 public class FanearkidValiderer {
     public static final String KONFIG_FANEARKID = "fanearkid";
 
-    public static Valideringsfeil valider(PersondataDTO dto, KonfigparamTjeneste konfig){
-
+    public static Valideringsfeil valider(PersondataDTO dto, KonfigparamTjeneste konfig) {
         Integer fieldLength = konfig.getInt(KONFIG_FANEARKID);
         String fanearkid = dto.getFanearkid();
-        if (fieldLength!=null){
-            if (fanearkid==null){
-                return new Valideringsfeil("fanearkid", "FeilFanearkidNull", "{fanearkidfeltlengde :"+fieldLength.toString()+"}");
+        if (fieldLength != null) {
+            if (fanearkid == null) {
+                return new Valideringsfeil("fanearkid", "FeilFanearkidNull", "{fanearkidfeltlengde :" + fieldLength.toString() + "}");
             }
-            if (fanearkid.length()!=fieldLength.intValue()){
-                return new Valideringsfeil("fanearkid", "FeilFanearkid", "{fanearkidfeltlengde :"+fieldLength.toString()+"}");
+            if (fanearkid.length() != fieldLength) {
+                return new Valideringsfeil("fanearkid", "FeilFanearkid", "{fanearkidfeltlengde :" + fieldLength.toString() + "}");
             }
         }
+        
         return null;
     }
 }
